requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
openai
azure-cosmos
google-generativeai
